Complete Golang API client for Divoom PIXOO64 LED display device.
  Features:
  - One-line text display with DisplayText() helper
  - Custom text positioning, fonts, colors, and alignment
  - GIF animation support (upload custom or play from URL)
  - Full device control (brightness, rotation, channels, tools)
  - Channel management with CustomPageIndex support
  - Helper methods for common operations
  - Zero external dependencies (stdlib only)

  Key Implementation Details:
  - Discovered critical requirement: ResetGifID() before uploads
  - Text requires Custom channel (3) + CustomPageIndex 1
  - Text overlays on animations (2-frame GIFs with different frames)
  - Proper timing delays for reliable operation

  Documentation:
  - Comprehensive README with examples
  - Interactive Swagger UI with OpenAPI 3.0 spec
  - Beginner-friendly getting started guide
  - Complete API reference
  - Technical solution documentation
  - Working examples verified on real hardware

  Examples:
  - Simple one-line text display
  - Custom positioned text with options
  - Colored backgrounds with text overlay
  - GIF from URL with text overlay

  All 4 examples tested and verified working on PIXOO64 device.

+++ types.go
@@ -74,9 +74,9 @@ type AllConfigResponse struct {
 type WeatherInfo struct {
 	ErrorCode  int     `json:"error_code"`
 	Weather    string  `json:"Weather"`
-	CurTemp    int     `json:"CurTemp"`
-	MinTemp    int     `json:"MinTemp"`
-	MaxTemp    int     `json:"MaxTemp"`
+	CurTemp    float64 `json:"CurTemp"`
+	MinTemp    float64 `json:"MinTemp"`
+	MaxTemp    float64 `json:"MaxTemp"`
 	Pressure   int     `json:"Pressure"`
 	Humidity   int     `json:"Humidity"`
 	Visibility int     `json:"Visibility"`
